Currently `cluster-api-provider-gcp` uses `sigs.k8s.io/controller-runtime v0.16.5` whereas `cluster-api-provider-aws` uses `sigs.k8s.io/controller-runtime v0.11.2`
In `machine-lifecycle-controller`, when we use both `cluster-api-provider-gcp` and `cluster-api-provider-aws`, the mismatch in `sigs.k8s.io/controller-runtime` is resulting in some missing packages when resolving the GO module file. A simpler fix was to remove the dependency of the `sigs.k8s.io/controller-runtime/pkg/metrics/server` package in `cluster-api-provider-gcp` as it was the least intensive approach from the code-refactoring perspective.